requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
